+++ TOMB4/specific/specificfx.cpp
@@ -1137,7 +1137,7 @@ void DrawDebris()
 		v[1].specular |= 0xFF000000;
 		v[2].specular |= 0xFF000000;
 
-		tex = &textinfo[(long)dptr->TextInfo & 0x7FFF];
+		tex = &textinfo[size_t(dptr->TextInfo) & 0x7FFF];
 		drawbak = tex->drawtype;
 
 		if (dptr->flags & 1)
