+++ TOMB4/tomb4/mod_config.cpp
@@ -206,7 +206,7 @@ void setup_custom_slots_for_level(int level, OBJECT_INFO* current_object_info_ar
             if (game_mod_config.level_info[level].objects_info.slot_override[i] != i) {
                 short original_nmeshes = current_object_info_array[i].nmeshes;
                 short original_mesh_index = current_object_info_array[i].mesh_index;
-                short original_bone_index = current_object_info_array[i].bone_index;
+                long original_bone_index = current_object_info_array[i].bone_index;
                 short *original_frame_base = current_object_info_array[i].frame_base;
                 short original_anim_index = current_object_info_array[i].anim_index;
